requirements.txt
asgiref==3.8.1
certifi==2024.2.2
cfgv==3.4.0
charset-normalizer==3.3.2
coverage==7.5.0
distlib==0.3.8
Django==4.2.11
djangorestframework==3.15.1
filelock==3.13.4
flake8==7.0.0
identify==2.5.36
idna==3.7
mccabe==0.7.0
nodeenv==1.8.0
platformdirs==4.2.1
pre-commit==3.7.0
pycodestyle==2.11.1
pyflakes==3.2.0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
sqlparse==0.5.0
typing_extensions==4.11.0
urllib3==2.2.1
virtualenv==20.26.0
